// packages/runner-ts/src/proto/intermediate.ts
export interface Reader {
  /** Channel URI. */
  uri: string;
}

/**
 * A RDFC channel writer.
 * Note that this is a declarative type, and must be instantiated by the runner.
 */
export interface Writer {
  /** Channel URI. */
  uri: string;
}
// ...
export interface ArgumentLiteral_List {
  values: ArgumentLiteral[];
}

export interface ArgumentMap {
  values: { [key: string]: Argument };
}

export interface ArgumentMap_ValuesEntry {
  key: string;
  value: Argument | undefined;
}

export interface ArgumentMap_List {
  values: ArgumentMap[];
}

/** An argument is either a literal or a map, or a list of literals or a list of maps. */
export interface Argument {
  literal?: ArgumentLiteral | undefined;
  literals?: ArgumentLiteral_List | undefined;
  map?: ArgumentMap | undefined;
  maps?: ArgumentMap_List | undefined;
}

/** A processor declaration. */
export interface Processor {
  /** The URI of the processor. */
  uri: string;
  /** The source code entrypoint. */
  entrypoint: string;
  /** Additional metadata, as required by the runners. */
  metadata: { [key: string]: string };
}

export interface Processor_MetadataEntry {
  key: string;
  value: string;
}

/** A concrete stage of a given processor, with instantiated arguments. */
export interface Stage {
  /** The URI of the stage. */
  uri: string;
  /** The processor definition. */
  processor: Processor | undefined;
  /** Stage arguments. */
  arguments: { [key: string]: Argument };
}

export interface Stage_ArgumentsEntry {
  key: string;
  value: Argument | undefined;
}
// ...
export const Argument = {
  encode(
    message: Argument,
    writer: _m0.Writer = _m0.Writer.create(),
  ): _m0.Writer {
    if (message.literal !== undefined) {
      ArgumentLiteral.encode(
        message.literal,
        writer.uint32(10).fork(),
      ).ldelim();
    }
    if (message.literals !== undefined) {
      ArgumentLiteral_List.encode(
        message.literals,
        writer.uint32(18).fork(),
      ).ldelim();
    }
    if (message.map !== undefined) {
      ArgumentMap.encode(message.map, writer.uint32(26).fork()).ldelim();
    }
    if (message.maps !== undefined) {
      ArgumentMap_List.encode(message.maps, writer.uint32(34).fork()).ldelim();
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): Argument {
    const reader =
      input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseArgument();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.literal = ArgumentLiteral.decode(reader, reader.uint32());
          continue;
        case 2:
          if (tag !== 18) {
            break;
          }

          message.literals = ArgumentLiteral_List.decode(
            reader,
            reader.uint32(),
          );
          continue;
        case 3:
          if (tag !== 26) {
            break;
          }

          message.map = ArgumentMap.decode(reader, reader.uint32());
          continue;
        case 4:
          if (tag !== 34) {
            break;
          }

          message.maps = ArgumentMap_List.decode(reader, reader.uint32());
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Argument {
    return {
      literal: isSet(object.literal)
        ? ArgumentLiteral.fromJSON(object.literal)
        : undefined,
      literals: isSet(object.literals)
        ? ArgumentLiteral_List.fromJSON(object.literals)
        : undefined,
      map: isSet(object.map) ? ArgumentMap.fromJSON(object.map) : undefined,
      maps: isSet(object.maps)
        ? ArgumentMap_List.fromJSON(object.maps)
        : undefined,
    };
  },

  toJSON(message: Argument): unknown {
    const obj: any = {};
    if (message.literal !== undefined) {
      obj.literal = ArgumentLiteral.toJSON(message.literal);
    }
    if (message.literals !== undefined) {
      obj.literals = ArgumentLiteral_List.toJSON(message.literals);
    }
    if (message.map !== undefined) {
      obj.map = ArgumentMap.toJSON(message.map);
    }
    if (message.maps !== undefined) {
      obj.maps = ArgumentMap_List.toJSON(message.maps);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Argument>, I>>(base?: I): Argument {
    return Argument.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Argument>, I>>(object: I): Argument {
    const message = createBaseArgument();
    message.literal =
      object.literal !== undefined && object.literal !== null
        ? ArgumentLiteral.fromPartial(object.literal)
        : undefined;
    message.literals =
      object.literals !== undefined && object.literals !== null
        ? ArgumentLiteral_List.fromPartial(object.literals)
        : undefined;
    message.map =
      object.map !== undefined && object.map !== null
        ? ArgumentMap.fromPartial(object.map)
        : undefined;
    message.maps =
      object.maps !== undefined && object.maps !== null
        ? ArgumentMap_List.fromPartial(object.maps)
        : undefined;
    return message;
  },
};

function createBaseProcessor(): Processor {
  return { uri: "", entrypoint: "", metadata: {} };
}

export const Processor = {
  encode(
    message: Processor,
    writer: _m0.Writer = _m0.Writer.create(),
  ): _m0.Writer {
    if (message.uri !== "") {
      writer.uint32(10).string(message.uri);
    }
    if (message.entrypoint !== "") {
      writer.uint32(18).string(message.entrypoint);
    }
    Object.entries(message.metadata).forEach(([key, value]) => {
      Processor_MetadataEntry.encode(
        { key: key as any, value },
        writer.uint32(34).fork(),
      ).ldelim();
    });
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): Processor {
    const reader =
      input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseProcessor();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.uri = reader.string();
          continue;
        case 2:
          if (tag !== 18) {
            break;
          }

          message.entrypoint = reader.string();
          continue;
        case 4:
          if (tag !== 34) {
            break;
          }

          const entry4 = Processor_MetadataEntry.decode(
            reader,
            reader.uint32(),
          );
          if (entry4.value !== undefined) {
            message.metadata[entry4.key] = entry4.value;
          }
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Processor {
    return {
      uri: isSet(object.uri) ? globalThis.String(object.uri) : "",
      entrypoint: isSet(object.entrypoint)
        ? globalThis.String(object.entrypoint)
        : "",
      metadata: isObject(object.metadata)
        ? Object.entries(object.metadata).reduce<{ [key: string]: string }>(
            (acc, [key, value]) => {
              acc[key] = String(value);
              return acc;
            },
            {},
          )
        : {},
    };
  },

  toJSON(message: Processor): unknown {
    const obj: any = {};
    if (message.uri !== "") {
      obj.uri = message.uri;
    }
    if (message.entrypoint !== "") {
      obj.entrypoint = message.entrypoint;
    }
    if (message.metadata) {
      const entries = Object.entries(message.metadata);
      if (entries.length > 0) {
        obj.metadata = {};
        entries.forEach(([k, v]) => {
          obj.metadata[k] = v;
        });
      }
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Processor>, I>>(base?: I): Processor {
    return Processor.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Processor>, I>>(
    object: I,
  ): Processor {
    const message = createBaseProcessor();
    message.uri = object.uri ?? "";
    message.entrypoint = object.entrypoint ?? "";
    message.metadata = Object.entries(object.metadata ?? {}).reduce<{
      [key: string]: string;
    }>((acc, [key, value]) => {
      if (value !== undefined) {
        acc[key] = globalThis.String(value);
      }
      return acc;
    }, {});
    return message;
  },
};
// ...
export const Stage = {
  encode(message: Stage, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.uri !== "") {
      writer.uint32(10).string(message.uri);
    }
    if (message.processor !== undefined) {
      Processor.encode(message.processor, writer.uint32(18).fork()).ldelim();
    }
    Object.entries(message.arguments).forEach(([key, value]) => {
      Stage_ArgumentsEntry.encode(
        { key: key as any, value },
        writer.uint32(26).fork(),
      ).ldelim();
    });
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): Stage {
    const reader =
      input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseStage();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.uri = reader.string();
          continue;
        case 2:
          if (tag !== 18) {
            break;
          }

          message.processor = Processor.decode(reader, reader.uint32());
          continue;
        case 3:
          if (tag !== 26) {
            break;
          }

          const entry3 = Stage_ArgumentsEntry.decode(reader, reader.uint32());
          if (entry3.value !== undefined) {
            message.arguments[entry3.key] = entry3.value;
          }
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Stage {
    return {
      uri: isSet(object.uri) ? globalThis.String(object.uri) : "",
      processor: isSet(object.processor)
        ? Processor.fromJSON(object.processor)
        : undefined,
      arguments: isObject(object.arguments)
        ? Object.entries(object.arguments).reduce<{ [key: string]: Argument }>(
            (acc, [key, value]) => {
              acc[key] = Argument.fromJSON(value);
              return acc;
            },
            {},
          )
        : {},
    };
  },

  toJSON(message: Stage): unknown {
    const obj: any = {};
    if (message.uri !== "") {
      obj.uri = message.uri;
    }
    if (message.processor !== undefined) {
      obj.processor = Processor.toJSON(message.processor);
    }
    if (message.arguments) {
      const entries = Object.entries(message.arguments);
      if (entries.length > 0) {
        obj.arguments = {};
        entries.forEach(([k, v]) => {
          obj.arguments[k] = Argument.toJSON(v);
        });
      }
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Stage>, I>>(base?: I): Stage {
    return Stage.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Stage>, I>>(object: I): Stage {
    const message = createBaseStage();
    message.uri = object.uri ?? "";
    message.processor =
      object.processor !== undefined && object.processor !== null
        ? Processor.fromPartial(object.processor)
        : undefined;
    message.arguments = Object.entries(object.arguments ?? {}).reduce<{
      [key: string]: Argument;
    }>((acc, [key, value]) => {
      if (value !== undefined) {
        acc[key] = Argument.fromPartial(value);
      }
      return acc;
    }, {});
    return message;
  },
};
// ...
export type DeepPartial<T> = T extends Builtin
  ? T
  : T extends globalThis.Array<infer U>
    ? globalThis.Array<DeepPartial<U>>
    : T extends ReadonlyArray<infer U>
      ? ReadonlyArray<DeepPartial<U>>
      : T extends {}
        ? { [K in keyof T]?: DeepPartial<T[K]> }
        : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin
  ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & {
      [K in Exclude<keyof I, KeysOfUnion<P>>]: never;
    };